MVP Architecture in LoanAccountFragment
