+++ mifosng-android/src/main/java/com/mifos/mifosxdroid/online/loanaccount/LoanAccountMvpView.java
@@ -1,0 +1,20 @@
+package com.mifos.mifosxdroid.online.loanaccount;
+
+import com.mifos.mifosxdroid.base.MvpView;
+import com.mifos.objects.organisation.ProductLoans;
+
+import java.util.List;
+
+import retrofit.client.Response;
+
+/**
+ * Created by Rajan Maurya on 08/06/16.
+ */
+public interface LoanAccountMvpView extends MvpView {
+
+    void showAllLoan(List<ProductLoans> productLoanses);
+
+    void showLoanAccountTemplate(Response response);
+
+    void showFetchingError(String s);
+}
