+++ mifosng-android/src/main/java/com/mifos/mifosxdroid/online/loanaccount/LoanAccountPresenter.java
@@ -1,0 +1,92 @@
+package com.mifos.mifosxdroid.online.loanaccount;
+
+import com.mifos.api.DataManager;
+import com.mifos.mifosxdroid.base.BasePresenter;
+import com.mifos.objects.organisation.ProductLoans;
+
+import java.util.List;
+
+import retrofit.client.Response;
+import rx.Subscriber;
+import rx.Subscription;
+import rx.android.schedulers.AndroidSchedulers;
+import rx.schedulers.Schedulers;
+
+/**
+ * Created by Rajan Maurya on 08/06/16.
+ */
+public class LoanAccountPresenter extends BasePresenter<LoanAccountMvpView> {
+
+    private final DataManager mDataManager;
+    private Subscription mSubscription;
+
+    public LoanAccountPresenter(DataManager dataManager) {
+        mDataManager = dataManager;
+    }
+
+    @Override
+    public void attachView(LoanAccountMvpView mvpView) {
+        super.attachView(mvpView);
+    }
+
+    @Override
+    public void detachView() {
+        super.detachView();
+        if (mSubscription != null) mSubscription.unsubscribe();
+    }
+
+    public void loadAllLoans() {
+        checkViewAttached();
+        getMvpView().showProgressbar(true);
+        if (mSubscription != null) mSubscription.unsubscribe();
+        mSubscription = mDataManager.getAllLoans()
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribeOn(Schedulers.io())
+                .subscribe(new Subscriber<List<ProductLoans>>() {
+                    @Override
+                    public void onCompleted() {
+                        getMvpView().showProgressbar(false);
+                    }
+
+                    @Override
+                    public void onError(Throwable e) {
+                        getMvpView().showProgressbar(false);
+                        getMvpView().showFetchingError("Failed to loans");
+                    }
+
+                    @Override
+                    public void onNext(List<ProductLoans> productLoanses) {
+                        getMvpView().showProgressbar(false);
+                        getMvpView().showAllLoan(productLoanses);
+                    }
+                });
+    }
+
+    public void loadLoanAccountTemplate(int clientId, int productId) {
+        checkViewAttached();
+        getMvpView().showProgressbar(true);
+        if (mSubscription != null) mSubscription.unsubscribe();
+        mSubscription = mDataManager.getLoansAccountTemplate(clientId, productId)
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribeOn(Schedulers.io())
+                .subscribe(new Subscriber<Response>() {
+                    @Override
+                    public void onCompleted() {
+                        getMvpView().showProgressbar(false);
+                    }
+
+                    @Override
+                    public void onError(Throwable e) {
+                        getMvpView().showProgressbar(false);
+                        getMvpView().showFetchingError("");
+                    }
+
+                    @Override
+                    public void onNext(Response response) {
+                        getMvpView().showProgressbar(false);
+                        getMvpView().showLoanAccountTemplate(response);
+                    }
+                });
+    }
+
+}
